refactor(router): restructure utilities
