+++ packages/router/src/browser-viewer-store.ts
@@ -2,7 +2,7 @@ import { EventAggregator, IEventAggregator } from '@aurelia/kernel';
 import { IWindow, IHistory, ILocation, IPlatform } from '@aurelia/runtime-html';
 import { NavigatorStateChangeEvent } from './events.js';
 import { INavigatorState, INavigatorStore, INavigatorViewer, INavigatorViewerOptions, NavigatorViewerState } from './navigator.js';
-import { QueueTask, TaskQueue } from './task-queue.js';
+import { QueueTask, TaskQueue } from './utilities/task-queue.js';
 
 /**
  * @internal
