+++ packages/router/src/routing-scope.ts
@@ -12,12 +12,12 @@ import { IRouter } from './router.js';
 import { RoutingInstruction } from './instructions/routing-instruction.js';
 import { LoadInstructionResolver } from './type-resolvers.js';
 import { Viewport, IViewportOptions } from './viewport.js';
-import { arrayRemove } from './utils.js';
-import { Collection } from './collection.js';
+import { arrayRemove } from './utilities/utils.js';
+import { Collection } from './utilities/collection.js';
 import { IConfigurableRoute, RouteRecognizer } from './route-recognizer.js';
 import { Navigation } from './navigation.js';
 import { NavigationCoordinator } from './navigation-coordinator.js';
-import { Runner, Step } from './runner.js';
+import { Runner, Step } from './utilities/runner.js';
 import { IRoute, Route } from './route.js';
 import { IConnectedCustomElement } from './endpoints/endpoint.js';
 import { RouterOptions } from './router-options.js';
