+++ packages/router/src/state-coordinator.ts
@@ -6,7 +6,7 @@
  */
 /* eslint-disable @typescript-eslint/no-empty-function */
 import { Navigation } from './navigation.js';
-import { OpenPromise } from './open-promise.js';
+import { OpenPromise } from './utilities/open-promise.js';
 
 export class Entity<T, S> {
   public running: boolean = false;
