+++ packages/router/src/utilities/collection.ts
