+++ packages/router/src/utilities/readme.md
@@ -1,0 +1,1 @@
+This folder contains utilities; functionality and features that aren't directly connected to the router but stand on their own.
